breaking: force image_uri to be passed for legacy TF images
This change also reorganizes the TF unit tests a bit,
and updates the tf_version fixture to include recent versions.
