+++ tests/conftest.py
@@ -194,6 +194,21 @@ def xgboost_version(request):
         "1.9.0",
         "1.10",
         "1.10.0",
+        "1.11",
+        "1.11.0",
+        "1.12",
+        "1.12.0",
+        "1.13",
+        "1.14",
+        "1.14.0",
+        "1.15",
+        "1.15.0",
+        "1.15.2",
+        "2.0",
+        "2.0.0",
+        "2.0.1",
+        "2.1",
+        "2.1.0",
     ],
 )
 def tf_version(request):
